bug fix: overlapCount for number of genes between peak and target
overlapCount for number of genes between peak and target was not right for the peak which is in 3' UTR of the gene. By default it was also counting the gene/transcript for which 3UTR it overlaps to. So if there was any immediate downstream gene, it was not marked based on the condition that there should be no gene inbetween. Now using (tx - (5UTR + 3UTR)) region